pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "jukkabot"
version = "0.1.0"
description = "Prototype Discord music bot with per-server queueing"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord.py[voice]>=2.5.0",
    "python-dotenv>=1.0.0",
    "yt-dlp>=2025.1.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
